Renamed /devops to /devops_platform
